Squashed commit of the following:
commit 2156e1ec94b5cd5dfbc8b1e6a90b68e28d3c4008
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Tue Mar 8 12:04:43 2016 -0300

    INT-8402: Minor cleanup 4

commit 6bd6f4f65829cde28196cffc66a63b3c082eeeab
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Tue Mar 8 11:46:05 2016 -0300

    INT-8402: Minor cleanup3

commit 90dc063437260fcfbfe3b47d9f45cb8dcbb6bdfb
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Tue Mar 8 11:26:09 2016 -0300

    INT-8402: Minor cleanup2

commit b1defd4bf715ace25323f70402894b8cba6f323d
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Tue Mar 8 11:13:24 2016 -0300

    INT-8402: Minor cleanup

commit 9cf8072c8018d7bbe61cdcc9b7aeb3927c878b6d
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Tue Mar 8 10:59:48 2016 -0300

    INT-8402: Refactored fixture loading, updated headline data

commit 5053416d4f3a79a666414674232cfe40eb6275db
Merge: 811b1d1 7ee9cdb
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Tue Mar 8 10:09:22 2016 -0300

    Merge remote-tracking branch 'origin/release/2.9+2' into update/INT-8402_unittestSuite

commit 811b1d14fbf06e9b93f7c3661997dce3140b72cb
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Thu Mar 3 11:51:16 2016 -0300

    INT-8402: Refactored fixture loading

commit 8ce8e8f94bff00d870a7f68c3347e62b3126fbb4
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Wed Mar 2 16:45:44 2016 -0300

    INT-8402: Silence one warning

commit 7c96a147f179435318c9ea5179809fb6f227280f
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Wed Mar 2 16:13:52 2016 -0300

    INT-8402: added missing file

commit b6027443921d130e99f587a18ec0b1f1d20830e6
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Wed Mar 2 15:29:18 2016 -0300

    INT-8402: refactoring according to the code review comments

commit 0cce8557d162e57c4ad6d0d515f9fff480e4bc73
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Mon Feb 29 11:32:09 2016 -0300

    INT-8402: cleanup

commit 7706006bc501be9f94a701cdd80768d3c0be26fd
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Mon Feb 29 11:21:37 2016 -0300

    INT-8402: cleanup

commit e539be4762587c0de6b2318f5a5818d13129be58
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Fri Feb 26 13:45:25 2016 -0300

    INT-8402: Another cleanup and schema tests update

commit ad7b14e91c9ec24a74914ed0e2d76e5e59f16d4f
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Fri Feb 26 13:12:18 2016 -0300

    INT-8402: Updated tests, added first schema validation tests

commit b3530883a74e668ffddaef3afd58a498c44797f9
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Fri Feb 26 11:45:32 2016 -0300

    INT-8402: updated cache

commit 6996a26101f11f6509c3772711994003962fca22
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Fri Feb 26 11:43:03 2016 -0300

    INT-8402: Minor cleanup and updated thirpartylib info

commit ccc83eb76fe7642a668a22c9248b56f9c347fb99
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Thu Feb 25 00:30:05 2016 -0300

    INT-8402: Description of the test faliure improved

commit 161a4d2d6c7da207fa6c47c28066be224eb76aaf
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Thu Feb 25 00:28:40 2016 -0300

    INT-8402: additional change

commit f1abe1d39713cc353008be497d8ec05f1f87cee7
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Thu Feb 25 00:17:47 2016 -0300

    INT-8402: Minor change

commit d95943d6347de67610e6c259eb8af27717c3740f
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Thu Feb 25 00:13:17 2016 -0300

    INT-8402: Added json-schema library and basic validation support

commit 30e1bf43ee7243919a660866cf0e0f8281462ff3
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Wed Feb 24 23:17:55 2016 -0300

    INT-8402: Added first login test

commit fef8c0925aed74fa6fa1d2442cca607c8148df3d
Merge: ab6a845 cbb2f32
Author: Darko Miletic <darko.miletic@gmail.com>
Date:   Mon Feb 22 15:41:32 2016 -0300

    Merge remote-tracking branch 'origin/release/2.9+2' into update/INT-8402_unittestSuite

commit ab6a845bea94726622da1988188eb1eda515b715
Merge: 2b9e6f0 fbe5397
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Mon Jan 25 10:21:22 2016 -0300

    Merge remote-tracking branch 'upstream/maint' into update/INT-8402_unittestSuite

commit 2b9e6f0faa788cb7aeccb305d7a39d500e457bdc
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Thu Jan 14 11:12:52 2016 -0300

    INT-8402: Sample test 3

commit ce0f704b298bc37dc40d457dd9fe308af7812879
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Thu Jan 14 11:06:09 2016 -0300

    INT-8402: Sample test 2

commit de07f39a8b7a081db7bc9f0b5aaf88c525e9c7a2
Author: Darko Miletic <darko.miletic@blackboard.com>
Date:   Wed Jan 13 18:36:00 2016 -0300

    INT-8402: Sample test

+++ classes/local/api/cache.php
@@ -44,6 +44,9 @@ class cache {
     protected $muc = null;
 
     public function __construct() {
+        if (PHPUNIT_TEST) {
+            return;
+        }
         $this->muc = \cache::make('local_xray', 'request');
     }
 
@@ -91,6 +94,13 @@ class cache {
      * @return bool|string
      */
     public function get($param) {
+        if (PHPUNIT_TEST) {
+            $result = testhelper::get_fixture_data($param);
+            if ($result) {
+                xrayws::instance()->setcookie('connect.sid=somecookievalue');
+            }
+            return $result;
+        }
         $result = false;
         if (self::cache_timeout() == 0) {
             return $result;
@@ -116,6 +126,9 @@ class cache {
      * @return void
      */
     public function set($param, $val) {
+        if (PHPUNIT_TEST) {
+            return;
+        }
         if (self::cache_timeout() > 0) {
             list($key, $created) = $this->getkeys($param);
             $this->muc->set_many([$key => serialize($val), $created => time()]);
@@ -127,6 +140,9 @@ class cache {
      * @return int
      */
     public function delete($param) {
+        if (PHPUNIT_TEST) {
+            return 0;
+        }
         return $this->muc->delete_many($this->getkeys($param));
     }
 
@@ -134,6 +150,9 @@ class cache {
      * @return bool
      */
     public function refresh() {
+        if (PHPUNIT_TEST) {
+            return false;
+        }
         return $this->muc->purge();
     }
 }
